LayoutTests/fast/events/page-scaled-mouse-click-expected.txt

This tests that page scaling does not affect mouse event pageX and pageY coordinates.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS eventLog is "click(10, 10)"
PASS eventLog is "click(20, 20)"
PASS successfullyParsed is true

TEST COMPLETE
